Add eureka client to notification service
