+++ notification-service/Dockerfile
@@ -21,4 +21,4 @@ ARG DEPENDENCY=app/target/dependency
 COPY --from=build ${DEPENDENCY}/BOOT-INF/lib /app/lib
 COPY --from=build ${DEPENDENCY}/META-INF /app/META-INF
 COPY --from=build ${DEPENDENCY}/BOOT-INF/classes /app
-ENTRYPOINT ["java", "-cp","app:app/lib/*","com.eirsteir.coffeewithme.notificationapi.NotificationApiApplication" ]
+ENTRYPOINT ["java", "-cp","app:app/lib/*","com.eirsteir.coffeewithme.notification.NotificationApiApplication" ]
