+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/NotificationServiceApplication.java
@@ -1,0 +1,16 @@
+package com.eirsteir.coffeewithme.notification;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+
+@SpringBootApplication
+@EnableEurekaClient
+public class NotificationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NotificationServiceApplication.class, args);
+    }
+
+}
