+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/config/WebSocketConfig.java
@@ -1,0 +1,17 @@
+package com.eirsteir.coffeewithme.notification.config;
+
+
+import org.springframework.stereotype.Component;
+import org.springframework.web.socket.config.annotation.AbstractWebSocketMessageBrokerConfigurer;
+import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
+import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
+
+@Component
+@EnableWebSocketMessageBroker
+public class WebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
+
+    @Override
+    public void registerStompEndpoints(StompEndpointRegistry registry) {
+        registry.addEndpoint("/ws").setAllowedOrigins("*").withSockJS(); // "http://client:3000"
+    }
+}
