+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/domain/NotificationType.java
@@ -1,0 +1,8 @@
+package com.eirsteir.coffeewithme.notification.domain;
+
+public enum NotificationType {
+
+    FRIENDSHIP_REQUESTED,
+    FRIENDSHIP_ACCEPTED
+
+}
