+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/dto/NotificationDto.java
@@ -1,0 +1,26 @@
+package com.eirsteir.coffeewithme.notification.dto;
+
+import com.eirsteir.coffeewithme.notification.domain.Notification;
+import com.eirsteir.coffeewithme.notification.domain.NotificationType;
+import lombok.*;
+import lombok.experimental.Accessors;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@ToString
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+@Accessors(chain = true)
+public class NotificationDto {
+
+    private Long notificationId;
+    private Date timestamp;
+    private Notification.UserDetails user;
+    private Boolean seen;
+    private NotificationType type;
+
+}
