+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/exception/EntityType.java
@@ -1,0 +1,7 @@
+package com.eirsteir.coffeewithme.notification.exception;
+
+public enum EntityType {
+
+    NOTIFICATION
+
+}
