--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/NotificationApiApplication.java
@@ -1,14 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class NotificationApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NotificationApiApplication.class, args);
-    }
-
-}
