--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/config/ModelMapperConfig.java
@@ -1,20 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        ModelMapper modelMapper = new ModelMapper();
-        modelMapper.getConfiguration()
-                .setFieldMatchingEnabled(true)
-                .setFieldAccessLevel(org.modelmapper.config.Configuration.AccessLevel.PRIVATE);
-
-        return modelMapper;
-    }
-
-}
