--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/config/PropertiesConfig.java
@@ -1,18 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
-import org.springframework.stereotype.Component;
-
-@Component
-@PropertySource("classpath:exception.properties")
-public class PropertiesConfig {
-
-    @Autowired
-    private Environment env;
-
-    public String getConfigValue(String configKey) {
-        return env.getProperty(configKey);
-    }
-}
