--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/config/WebSocketConfig.java
@@ -1,17 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.config;
-
-
-import org.springframework.stereotype.Component;
-import org.springframework.web.socket.config.annotation.AbstractWebSocketMessageBrokerConfigurer;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-
-@Component
-@EnableWebSocketMessageBroker
-public class WebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/ws").setAllowedOrigins("*").withSockJS(); // "http://client:3000"
-    }
-}
