--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/domain/NotificationType.java
@@ -1,8 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.domain;
-
-public enum NotificationType {
-
-    FRIENDSHIP_REQUESTED,
-    FRIENDSHIP_ACCEPTED
-
-}
