--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/dto/NotificationDto.java
@@ -1,26 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.dto;
-
-import com.eirsteir.coffeewithme.notificationapi.domain.Notification;
-import com.eirsteir.coffeewithme.notificationapi.domain.NotificationType;
-import lombok.*;
-import lombok.experimental.Accessors;
-
-import java.util.Date;
-
-@Getter
-@Setter
-@ToString
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-@Accessors(chain = true)
-public class NotificationDto {
-
-    private Long notificationId;
-    private Date timestamp;
-    private Notification.UserDetails user;
-    private Boolean seen;
-    private NotificationType type;
-
-}
