--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/exception/EntityType.java
@@ -1,7 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.exception;
-
-public enum EntityType {
-
-    NOTIFICATION
-
-}
