--- notification-service/src/main/java/com/eirsteir/coffeewithme/notificationapi/web/request/NotificationRequest.java
@@ -1,15 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi.web.request;
-
-import com.eirsteir.coffeewithme.notificationapi.domain.NotificationType;
-import lombok.Getter;
-
-@Getter
-public class NotificationRequest {
-
-    private Long subjectId;
-    private NotificationType type;
-    private Long userId;
-    private String name;
-    private String username;
-
-}
