+++ notification-service/src/main/resources/application.properties
@@ -1,3 +1,5 @@
 server.port=9000
-spring.application.name=notifications
+spring.application.name=notification-service
 spring.profiles.active=@spring.profiles.active@
+
+eureka.client.service-url.default-zone=http://localhost:8761/eureka
