+++ notification-service/src/test/java/com/eirsteir/coffeewithme/notification/NotificationServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.eirsteir.coffeewithme.notification;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotificationServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
