+++ notification-service/src/test/java/com/eirsteir/coffeewithme/notification/testutil/JSONUtils.java
@@ -1,0 +1,12 @@
+package com.eirsteir.coffeewithme.notification.testutil;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class JSONUtils {
+
+    public static String asJsonString(Object object) throws JsonProcessingException {
+        ObjectMapper objectMapper = new ObjectMapper();
+        return objectMapper.writeValueAsString(object);
+    }
+}
