--- notification-service/src/test/java/com/eirsteir/coffeewithme/notificationapi/NotificationApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.eirsteir.coffeewithme.notificationapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NotificationApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
