+++ social-service/src/main/resources/application.properties
@@ -5,12 +5,6 @@ spring.profiles.active=@spring.profiles.active@
 
 eureka.client.service-url.default-zone=http://localhost:8761/eureka
 
-
-logging.level.web=DEBUG
-logging.level.com.eirsteir=DEBUG
-logging.level.org.springframework.security=DEBUG
-logging.level.org.springframework.web.servlet.mvc.method.annotation.ExceptionHandlerExceptionResolver=DEBUG
-
 spring.security.oauth2.client.registration.google.client-id=903677381406-jeb48ctobmujtje4h0mdpfvqir2n6s27.apps.googleusercontent.com
 spring.security.oauth2.client.registration.google.client-secret=8eGSCbAiW00rxnAQhjEEGq2z
